pyproject.toml
[project]
name = "nonebot-plugin-flo-jav"
version = "0.1.0"
description = "nonebot-plugin-flo-jav"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "nonebot2[fastapi,httpx,websockets]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "nb-cli>=1.5.0",
    "pip>=25.3",
    "nonebot-plugin-localstore>=0.7.4",
    "nonebot-plugin-uninfo>=0.9.0",
    "nonebot-plugin-alconna>=0.59.4",
    "curl-cffi>=0.13.0",
]

[tool.nonebot]
plugin_dirs = []
builtin_plugins = ["echo"]

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
nonebot-plugin-flo-jav = ["nonebot_plugin_flo_jav"]
nonebot-plugin-localstore = ["nonebot_plugin_localstore"]
nonebot-plugin-alconna = ["nonebot_plugin_alconna"]
nonebot-plugin-uninfo = ["nonebot_plugin_uninfo"]
